use event in_language field, refactor language selection
