+++ frontend/tet/admin/src/components/editor/contactPerson/ContactPerson.tsx
@@ -14,13 +14,7 @@ const ContactPerson: React.FC = () => {
   const { t } = useTranslation();
   const { required, name, email, phone } = useValidationRules();
   const theme = useTheme();
-  const languageOptions = [
-    { value: 'fi', label: t('common:editor.posting.contactLanguageFi') },
-    { value: 'sv', label: t('common:editor.posting.contactLanguageSv') },
-    { value: 'en', label: t('common:editor.posting.contactLanguageEn') },
-  ];
 
-  console.log(required);
   return (
     <FormSection header={t('common:editor.posting.header')}>
       <$GridCell
@@ -69,19 +63,7 @@ const ContactPerson: React.FC = () => {
         css={`
           row-gap: ${theme.spacing.xl};
         `}
-      >
-        <$GridCell $colSpan={3}>
-          <Dropdown
-            id="contact_language"
-            options={languageOptions}
-            initialValue={languageOptions[0]}
-            label={t('common:editor.posting.contactLanguage')}
-            registerOptions={{
-              required: required,
-            }}
-          />
-        </$GridCell>
-      </$GridCell>
+      ></$GridCell>
     </FormSection>
   );
 };
